♻️ refactor: 지원 상세 모달 관련 쿼리 로직을 커스텀 훅으로 리팩토링
