+++ src/features/application/index.ts
@@ -1,3 +1,4 @@
+export { useApplicationModalProps } from "./hooks/useApplicationModalProps"
 export { ApplicantDetailRow } from "./ui/ApplicantDetailRow"
 export { ApplicantTableHead } from "./ui/ApplicantTableHead"
 export { ApplicationDetailModal } from "./ui/ApplicationDetailModal"
